lists layout improved
