+++ src/scss/index.scss
@@ -34,7 +34,8 @@ hr {
   border: 0;
 }
 input[type="button"],
-input[type="radio"] {
+input[type="radio"],
+input[type="submit"] {
   display: none;
 }
 /* Change the white to any color */
